# #1745 - Use SingleEntryVector in box overapproximation