@mforets 2020-06-02
> Ball1 and Ball2 (slower)

for Ball1, i would expect `box_approximation` to be a no-op? for `Ball2`, i guess we dispatch on the SEV