@schillic 2020-06-02
> for Ball1, i would expect box_approximation to be a no-op?

Do you mean `BallInf`?

> for Ball2, i guess we can dispatch on the SEV?

Yes, in the support function. But currently we do not do that (in fact for none of the set types).