@mforets 2020-06-02
oh, true i read `BallInf` :) 

there's dispatch on SEV but only for a small number of sets (See https://github.com/JuliaReach/LazySets.jl/projects/3 and https://github.com/JuliaReach/LazySets.jl/pull/1785 for hyperrectangular sets and SIH)