@schillic 2020-06-02
> for hyperrectangular sets and SIH

Yes, but for those the `box_approximation` is anyway trivial.